+++ dflib/src/main/java/com/nhl/dflib/ColumnDataFrame.java
@@ -306,12 +306,6 @@ public class ColumnDataFrame implements DataFrame {
         return new DataFrameSorter(this).sort(sorters);
     }
 
-    @Deprecated
-    @Override
-    public <V extends Comparable<? super V>> DataFrame sort(RowToValueMapper<V> sortKeyExtractor) {
-        return new DataFrameSorter(this).sort(sortKeyExtractor);
-    }
-
     @Override
     public DataFrame sort(String[] columns, boolean[] ascending) {
         return new DataFrameSorter(this).sort(columns, ascending);