+++ dflib/src/main/java/com/nhl/dflib/DataFrame.java
@@ -640,12 +640,6 @@ public interface DataFrame extends Iterable<RowProxy> {
      */
     DataFrame sort(Sorter... sorters);
 
-    /**
-     * @deprecated since 0.12 as sorting by RowToValueMapper is redundant, and can be expressed as a Sorter.
-     */
-    @Deprecated
-    <V extends Comparable<? super V>> DataFrame sort(RowToValueMapper<V> sortKeyExtractor);
-
     DataFrame sort(String column, boolean ascending);
 
     DataFrame sort(int column, boolean ascending);