+++ dflib/src/main/java/com/nhl/dflib/RowToValueMapper.java
@@ -5,21 +5,5 @@ import com.nhl.dflib.row.RowProxy;
 @FunctionalInterface
 public interface RowToValueMapper<V> {
 
-    /**
-     * @deprecated since 0.12 in favor of {@link Exp} column expressions
-     */
-    @Deprecated
-    static <V> RowToValueMapper<V> columnReader(String column) {
-        return r -> (V) r.get(column);
-    }
-
-    /**
-     * @deprecated since 0.12 in favor of {@link Exp} column expressions
-     */
-    @Deprecated
-    static <V> RowToValueMapper<V> columnReader(int column) {
-        return r -> (V) r.get(column);
-    }
-
     V map(RowProxy row);
 }